Changing .SAV file to .CSV
I changed the .SAV file from Pew Research Center to a .CSV that way is easier for us to analyze and use the dataset. The only problem I think I'm currently is when I use the df.info line the output is shorter than it normally is, if you have any suggestion feel free to tell me so I can improve it!