Finish Naive Bayes
I finished my Naive Bayes, turns out my problem was that some of the columns are object types and not float type so I had to change all of that. I plan to add more columns for X. If you have any ideas what columns is best to add, please feel free to tell me, but I plan to look at the pdf more to see if there's any important things that I've missed.